# docs/swagger.yaml
basePath: /api/v1
definitions:
  common.PageData:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  dto.ChangePasswordRequest:
    properties:
      newPassword:
        maxLength: 128
        minLength: 6
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
...
  dto.CreateUserRequest:
    properties:
      deptId:
        type: integer
      email:
        type: string
      nickname:
        maxLength: 64
        type: string
      password:
        maxLength: 128
        minLength: 6
        type: string
      phone:
        type: string
      postIds:
        items:
          type: integer
        type: array
      remark:
        maxLength: 500
        type: string
      roleIds:
        items:
          type: integer
        type: array
      status:
        enum:
        - 0
        - 1
        type: integer
      username:
        maxLength: 64
        minLength: 2
        type: string
    required:
    - deptId
    - password
    - username
    type: object
...
  dto.ResetPasswordRequest:
    properties:
      id:
        type: integer
      password:
        maxLength: 128
        minLength: 6
        type: string
    required:
    - id
    - password
    type: object
  dto.StatusRequest:
    properties:
      id:
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - id
    - status
    type: object
...
  dto.UpdateUserDeptRequest:
    properties:
      deptId:
        type: integer
      id:
        type: integer
    required:
    - deptId
    - id
    type: object
  dto.UpdateUserRequest:
    properties:
      deptId:
        type: integer
      email:
        type: string
      id:
        type: integer
      nickname:
        maxLength: 64
        type: string
      phone:
        type: string
      postIds:
        items:
          type: integer
        type: array
      remark:
        maxLength: 500
        type: string
      roleIds:
        items:
          type: integer
        type: array
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - id
    type: object
  dto.UpdateUserRolesRequest:
    properties:
      id:
        type: integer
      roleIds:
        items:
          type: integer
        type: array
    required:
    - id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Gin-Admin 后台管理系统 API 文档
  title: Gin-Admin API
  version: 1.0.0
paths:
...
  /api/v1/system/user:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 创建用户
      tags:
      - 管理员
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 更新用户
      tags:
      - 管理员
  /api/v1/system/user/{id}:
    delete:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 删除用户
      tags:
      - 管理员
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 获取用户详情
      tags:
      - 管理员
  /api/v1/system/user/changePwd:
    put:
      consumes:
      - application/json
      parameters:
      - description: 密码信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: 修改密码
      tags:
      - 管理员
  /api/v1/system/user/dept:
    put:
      consumes:
      - application/json
      parameters:
      - description: 部门
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserDeptRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 修改用户部门
      tags:
      - 管理员
  /api/v1/system/user/list:
    get:
      parameters:
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 手机号
        in: query
        name: phone
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      - description: 部门ID
        in: query
        name: deptId
        type: integer
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页条数
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/common.PageData'
              type: object
      summary: 用户列表
      tags:
      - 管理员
  /api/v1/system/user/resetPwd:
    put:
      consumes:
      - application/json
      parameters:
      - description: 新密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 重置密码
      tags:
      - 管理员
  /api/v1/system/user/roles:
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 修改用户角色
      tags:
      - 管理员
  /api/v1/system/user/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: 状态
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.StatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
      summary: 修改用户状态
      tags:
      - 管理员
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: '输入格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"